Gosh dang it, it works! Controlling 0-9 levels of resistance from wahoo fitness app using bluetooth.

+++ nordic/emotion/resistance.c
@@ -3,7 +3,7 @@
 
 bool m_initialized = false;
 
-static uint32_t RESISTANCE_LEVEL[MAX_RESISTANCE_LEVELS] = { 
+static uint16_t RESISTANCE_LEVEL[MAX_RESISTANCE_LEVELS] = { 
 	2107, // 0 - no resistance
 	1300,	// this was 1273
 	1272,	
@@ -21,7 +21,7 @@ static void init_resistance()
 	m_initialized = true;
 }
 
-void set_resistance(uint32_t level)
+void set_resistance(uint8_t level)
 {
 		if (!m_initialized) 
 			init_resistance();
