+++ nordic/emotion/resistance.h
@@ -6,6 +6,6 @@
 #define PIN_SERVO_SIGNAL		3		// P3 - P0.03
 #define MAX_RESISTANCE_LEVELS 10
 
-void set_resistance(uint32_t level);
+void set_resistance(uint8_t level);
 
 #endif
